% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chk_required_cols}
\alias{chk_required_cols}
\title{Assert that required columns exist}
\usage{
chk_required_cols(df, cols, df_name = deparse(substitute(df)))
}
\description{
Assert that required columns exist
}
\keyword{internal}
