% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_maps.R
\name{fdr_plot_downscaled_GHG}
\alias{fdr_plot_downscaled_GHG}
\title{Plot downscaled GHG emissions}
\usage{
fdr_plot_downscaled_GHG(
  out_res,
  rasterized_layer,
  ns_map,
  border_sf = NULL,
  year = NULL,
  limits = NULL,
  LU = NULL,
  na_color = "grey90",
  add_border = TRUE
)
}
\arguments{
\item{out_res}{results table (typically fdr_run_downscaling()$out.res or $downscaled_LUC)
must have: ns, lu.to, times, value}

\item{rasterized_layer}{SpatRaster with ns_int values (from fdr_build_id_maps())}

\item{ns_map}{id_c -> ns_int mapping (from fdr_build_ns_map())}

\item{na_color}{fill color for NA pixels}
}
\description{
Plot downscaled GHG emissions
}
