% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_validate_inputs.R
\name{fdr_validate_inputs}
\alias{fdr_validate_inputs}
\title{Validate required inputs are present and minimally well-formed}
\usage{
fdr_validate_inputs(inputs)
}
\arguments{
\item{inputs}{Output of fdr_load_inputs()}
}
\value{
TRUE invisibly
}
\description{
Validate required inputs are present and minimally well-formed
}
