% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_wrangle.R
\name{fdr_wrangle_fable_targets}
\alias{fdr_wrangle_fable_targets}
\title{Wrangle FABLE land transition matrix into DownscalR targets format}
\usage{
fdr_wrangle_fable_targets(fable_sheet, min_year = 2020)
}
\arguments{
\item{fable_sheet}{A data.frame/tibble like inputs$FABLE_targets (raw Excel sheet)}

\item{min_year}{Keep only targets with times >= min_year (default 2020)}
}
\value{
tibble with columns: times, lu.from, lu.to, value
}
\description{
Wrangle FABLE land transition matrix into DownscalR targets format
}
