% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_downscale.R
\name{fdr_run_downscaling}
\alias{fdr_run_downscaling}
\title{Fit MNL models and run DownscalR downscaling}
\usage{
fdr_run_downscaling(
  targets,
  country_luc,
  priors,
  mnl_niter = 100,
  mnl_nburn = 50
)
}
\arguments{
\item{targets}{FABLE transition targets (long: times, lu.from, lu.to, value)}

\item{country_luc}{Calibration transitions (output of lc_build_country_luc)}

\item{priors}{Output of fdr_build_priors()}

\item{mnl_niter}{Number of MCMC iterations}

\item{mnl_nburn}{Burn-in iterations}
}
\value{
List with:
\itemize{
\item downscaled_LUC
\item X_long
\item pred_coeff_long
\item country_start_areas
}
}
\description{
Fit MNL models and run DownscalR downscaling
}
